% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{soft_thresh}
\alias{soft_thresh}
\title{ST function}
\usage{
soft_thresh(a, lambda)
}
\arguments{
\item{a}{- aa}

\item{lambda}{- bb}
}
\value{
st value
}
\description{
ST function
}
\examples{
soft_thresh(3,1)
}
